+++ src/App.jsx
@@ -4,7 +4,7 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import * as bootstrap from "bootstrap";
 
 import Header from "./components/Header";
-import Home from "./components/Home/Home";
+import Home from "./components/Home";
 import About from "./components/About";
 import Contact from "./components/Contact";
 import Answers from "./components/Answers";
