+++ src/components/Home/index.jsx
@@ -2,7 +2,7 @@ import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Container from "react-bootstrap/esm/Container";
-import "./styles.css";
+import "./index.css";
 
 import OpenAI from "openai";
