+++ src/index.css
@@ -2,7 +2,7 @@
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
-  padding-top: 108px;
+  padding-top: 32px;
 
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
